WIP: Added main bottom nav controls
This fixes an issue with the menu structure generation, where as "deleted" jaen pages are still processed in the structure.
The main bottom nav now receives the first dynamic logic to automatically determine the next and previous page. Please note that this is heavily work in progress and does not work properly.