feat: Added code preview and playground component
This introduces two components that interplays closely with the improved CodeSnippet component.

CodePlayground:
This component consists of the existing CodeSnippet and the new CodeResultPreview component allowing to easily preview written code.

CodeResultPreview:
While it has a similar style to the CodeSnippet component, it's key purpose is to display certain results that can be any valid ReactNode.